requirements.txt
fastapi
uvicorn
starlette
line-bot-sdk
openai
